Changed setAreaInFeet to a Float instead of Int

+++ 2. Calculations/9. Paint Calculator/paint_calculator/src/PaintPerAreaCalculator.cpp
@@ -1,6 +1,6 @@
 #include "PaintPerAreaCalculator.h"
 
-void PaintPerAreaCalculator::setAreaInFeet(int areaInFeet)
+void PaintPerAreaCalculator::setAreaInFeet(float areaInFeet)
 {
 }
